karatep: init: remove unneeded '&' from the fscanf statement.
An array "decays" into a pointer to its first element, so fscanf(lenovo_id, "%s", board_id) is equivalent to fscanf(lenovo_id, "%s", &board_id). On the other hand, fscanf(lenovo_id, "%s", &board_id) passes a pointer-to-char[15], but it points to the same place.
So its safe to remove extra trailing '&'.